refactor(deploy): isolate extension policy loading
