+++ src/core/component/mod.rs
@@ -504,81 +504,6 @@ impl Component {
             }
         }
     }
-
-    pub fn deploy_protected_path_suffixes(&self) -> Vec<String> {
-        let mut suffixes = HashSet::new();
-        let Some(extensions) = self.extensions.as_ref() else {
-            return Vec::new();
-        };
-
-        for extension_id in extensions.keys() {
-            let Some(deploy_policy) = extension_deploy_policy(extension_id) else {
-                continue;
-            };
-
-            let Some(policy_suffixes) = deploy_policy
-                .get("protected_path_suffixes")
-                .and_then(serde_json::Value::as_array)
-            else {
-                continue;
-            };
-
-            for suffix in policy_suffixes.iter().filter_map(serde_json::Value::as_str) {
-                suffixes.insert(suffix.to_string());
-            }
-        }
-
-        let mut suffixes: Vec<String> = suffixes.into_iter().collect();
-        suffixes.sort();
-        suffixes
-    }
-
-    pub fn deploy_owner_hint_for_path(&self, remote_path: &str) -> Option<String> {
-        let extensions = self.extensions.as_ref()?;
-
-        for extension_id in extensions.keys() {
-            let Some(deploy_policy) = extension_deploy_policy(extension_id) else {
-                continue;
-            };
-
-            let Some(owner_hints) = deploy_policy
-                .get("owner_hints")
-                .and_then(serde_json::Value::as_array)
-            else {
-                continue;
-            };
-
-            for hint in owner_hints {
-                let Some(path_contains) = hint
-                    .get("path_contains")
-                    .and_then(serde_json::Value::as_str)
-                else {
-                    continue;
-                };
-                let Some(suggested_owner) = hint
-                    .get("suggested_owner")
-                    .and_then(serde_json::Value::as_str)
-                else {
-                    continue;
-                };
-
-                if remote_path.contains(path_contains) {
-                    return Some(suggested_owner.to_string());
-                }
-            }
-        }
-
-        None
-    }
-}
-
-fn extension_deploy_policy(
-    extension_id: &str,
-) -> Option<serde_json::Map<String, serde_json::Value>> {
-    let path = crate::paths::extension_manifest(extension_id).ok()?;
-    let raw = std::fs::read_to_string(path).ok()?;
-    let manifest: serde_json::Value = serde_json::from_str(&raw).ok()?;
-    manifest.get("deploy")?.as_object().cloned()
 }
 
 fn render_remote_path_template(template: &str, component_id: &str, dir_name: &str) -> String {
@@ -962,65 +887,6 @@ mod tests {
 
         component.resolve_remote_path();
         assert_eq!(component.remote_path, "custom/deploy/path");
-    }
-
-    #[test]
-    fn deploy_policy_collects_extension_protected_suffixes() {
-        with_isolated_home(|home| {
-            write_extension_fixture(
-                home,
-                "example",
-                r#"{
-    "protected_path_suffixes": ["/shared/plugins", "/shared/uploads"]
-  }"#,
-            );
-
-            let component = Component {
-                id: "my-plugin".to_string(),
-                extensions: Some(HashMap::from([(
-                    "example".to_string(),
-                    ScopedExtensionConfig::default(),
-                )])),
-                ..Component::default()
-            };
-
-            assert_eq!(
-                component.deploy_protected_path_suffixes(),
-                vec!["/shared/plugins", "/shared/uploads"]
-            );
-        });
-    }
-
-    #[test]
-    fn deploy_policy_resolves_owner_hint_for_matching_path() {
-        with_isolated_home(|home| {
-            write_extension_fixture(
-                home,
-                "example",
-                r#"{
-    "owner_hints": [
-      { "path_contains": "shared/plugins/", "suggested_owner": "www-data:www-data" }
-    ]
-  }"#,
-            );
-
-            let component = Component {
-                id: "my-plugin".to_string(),
-                extensions: Some(HashMap::from([(
-                    "example".to_string(),
-                    ScopedExtensionConfig::default(),
-                )])),
-                ..Component::default()
-            };
-
-            let hint = component
-                .deploy_owner_hint_for_path("shared/plugins/my-plugin")
-                .expect("owner hint");
-            assert_eq!(hint, "www-data:www-data");
-            assert!(component
-                .deploy_owner_hint_for_path("other/my-plugin")
-                .is_none());
-        });
     }
 
     // ========================================================================
